pyproject.toml
[tool.poetry]
name = "union for the protection of homeless animals"
version = "0.1.0"
description = "The Union for the Protection of Homeless Animals is an initiative dedicated to the welfare and protection of homeless animals. The organization brings together individuals committed to creating safe and caring conditions for animals that have found themselves without a home."
authors = ["Nickolaj <nikolajpetrishchev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
python-decouple = "^3.8"
pytest-asyncio = "^0.23.2"
pytest-asyncio-cooperative = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
